fixed player controller bound to card node
the card nodes were permanently bound to one of the player controllers; this means that when a card switched to the other player, in the view it was still considered a child of the original player. this fix corrects several bugs related to being unable to tap a card in a player's hand.